rename ./core.js to ./index.js
