+++ test/block3.js
@@ -17,7 +17,7 @@ var toAddress = require('../lib/util').toAddress
 // 2. alice never tries to connect to bob. (removed from peers)
 // 3. carol will not give bob any, she will not give him any data from alice.
 
-var createSbot = require('../core')
+var createSbot = require('../')
     .use(require('../plugins/friends'))
     .use(require('../plugins/replicate'))
     .use(require('../plugins/block'))
