+++ test/realtime.js
@@ -6,7 +6,7 @@ var pull      = require('pull-stream')
 
 var ssbKeys = require('ssb-keys')
 
-var createSbot = require('../core')
+var createSbot = require('../')
   .use(require('../plugins/friends'))
   .use(require('../plugins/replicate'))
   .use(require('../plugins/gossip'))
